x86_64: Check QEMU CPU state ELF note version and size
As of 2024, the only existing version is 1, but make the code future-proof
by ignoring any other version.

Also, check that the size stored in the state itself is not greater than
the size specified in the ELF note header. A smaller value is accepted,
because that might be due to padding (although no such data has been
observed in the wild).

Signed-off-by: Petr Tesarik <petr@tesarici.cz>