tests: Add root page table entries following x86_64 direct mapping
To initialize a translation system, the virtual address space may have to
be scanned for the direct mapping. The highest scanned address currently
corresponds to Linux version 2.6.32 memory layout, but test cases should
not rely on it by omitting data beyond that boundary.

Add zero entries up to the next mapping contained in the data file.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>