tests: distribute a Python script to extract page table hierarchy
This script can help create new test cases by extracting the complete page
table hierarchy from an existing dump file.

Note that the script does not add a dependency on Python or pykdumpfile,
because it is not used at build time. It is included here for people who
want to extend the test suite based on interesting dump file samples.

Since it is a development tool, the script should not be installed on the
target system.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>