x86_64: Handle dumps with non-empty LDT remap for PTI
After the direct mapping was moved by commit d52888aa2753 ("x86/mm: Move
LDT remap out of KASLR region on 5-level paging"), it no longer be the
first mapping encountered when scanning the Linux kernel virtual address
space. The LDT mapping may come first...

Since we don't know if there is an LDT mapping, we must verify that what
was found is indeed direct mapping. In fact, there is an is_directmap()
function already. It verifies whether the first address translates to
physical address 0. On x86 systems, it is reasonable to expect that there
is always RAM at physical address zero, because that's where interrupt
vectors are located in real mode, and every x86 CPU starts operating in
this mode after power-on or reset. It is also reasonable to expect that
nothing is ever mapped to that address except direct mapping (and possibly
the kernel text mapping), so the first kernel virtual address which maps to
physical address zero should mark the beginning of the direct mapping.

Add a test case based on a kernel booted with pti=on and running DOSEMU
(one of few applications with an LDT).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>