x86_64: Use Linux kernel page tables when PTI active and init from CR3
Switch to kernel page tables when:

- initializing from CR3 register value
- X86_FEATURE_PTI was active
- user-mode page tables were active

The switch itself is trivial (clear a bit in the root page table base
address), detecting when to switch is not. Kernels built without page table
isolation can allocate the root page table anywhere, including an address
where the corresponding bit is set, but in this case, the address should
not be altered.

One difference between kernel-mode page tables and user-mode page tables is
the absence of kernel direct mapping in user mode. Admittedly, a few pages
in the direct mapping are still present in user mode, because entry/exit
code may need to write into TSS, so the writable TSS alias must be
accessible, and this alias is direct-mapped. However, it can never be
located at physical address zero (e.g. because it is allocated with struct
cpu_entry_area but it is not the first member). Consequently, this part of
the direct mapping will not be recognized as such by libaddrxlat since
commit b08090e7bd85 ("x86_64: Handle dumps with non-empty LDT remap for
PTI").

Signed-off-by: Petr Tesarik <ptesarik@suse.com>