kdump_set_attr(): Always take ownership of the new value on error
Make sure that the caller does not leak the new attribute value if an error
occurs. Until now, the new value would be properly discarded on allocation
failures or if an error is returned by a pre-hook, but not when:

a. trying to set a non-existent attribute
b. trying to use the wrong value type

Fixes: #81
Signed-off-by: Petr Tesarik <ptesarik@suse.com>