Make kdump_get_typed_attr() easier to use
The current API is inconvenient. First, the caller must provide a
kdump_attr_t variable, where the type field is used only for input and the
value field is used only for output. Second, the type field must be set
before the call, which usually requires a separate statement. Third, the
variable is updated on type mismatch, even though the function returns an
error code.

The only reason for such clumsy API is that it was easy to implement.

To fix the issues:
- pass the attribute type and value parameters separately,
- do not change the output variable on type mismatch.

This is a BACKWARDS INCOMPATIBLE CHANGE, so reset libtool version.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>