Import kdumpid
Add the sources of kdumpid v1.6 (commit cb3d9452e2f6) and adapt them so
they can be built from libkdumpfile using GNU autotools.

The sources are identical, except these changes:
- include "config.h" from kdumpid.h
- remove some superfluous whitespace

The versioning scheme of kdumpid is retained for now to ensure that the
versions form a monotonically increasing sequence.

The original kdumpid repository will be set to read-only to avoid
potential divergence.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>